Added REST endpoint for winner fetching. Packaged everything in docker-compose.yml. Tests still failing
