+++ gradle.properties
@@ -8,3 +8,5 @@ jacksonVersion=2.14.2
 lombokVersion=1.18.24
 mysqlVersion=8.0.32
 h2Version=2.1.214
+lettuceVersion=6.2.2.RELEASE
+commonsPool2Version=2.11.1
