+++ lombok.config
@@ -1,4 +1,4 @@
 config.stopbubbling=true
 lombok.addLombokGeneratedAnnotation = true
-lombok.copyableannotations+=com.shankhadeepghoshal.springbatch.annotations.PersonImport
+lombok.copyableannotations+=com.shankhadeepghoshal.springbatch.personimport.PersonImport
 lombok.copyableannotations+=org.springframework.beans.factory.annotation.Value
