+++ pmd-ruleset.xml
@@ -2,8 +2,7 @@
 <ruleset name="DevFactory"
   xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0
-  http://pmd.sourceforge.net/ruleset_2_0_0.xsd">
+  xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 https://pmd.sourceforge.io/ruleset_2_0_0.xsd">
 
   <description>
     PMD Rules Configuration
@@ -52,19 +51,15 @@
   <rule ref="category/java/errorprone.xml"/>
   <rule ref="category/java/errorprone.xml/AvoidDuplicateLiterals">
     <properties>
-      <property name="separator" value="," />
-      <property name="maxDuplicateLiterals" value="2" />
-      <property name="minimumLength" value="3" />
-      <property name="skipAnnotations" value="false" />
-      <property name="exceptionList" value="" />
+      <property name="separator" value=","/>
+      <property name="maxDuplicateLiterals" value="2"/>
+      <property name="minimumLength" value="3"/>
+      <property name="skipAnnotations" value="false"/>
+      <property name="exceptionList" value=""/>
     </properties>
   </rule>
 
-  <rule ref="category/java/errorprone.xml/BeanMembersShouldSerialize">
-    <properties>
-      <property name="ignoredAnnotations" value="lombok.Data|lombok.Getter|lombok.Value"/>
-    </properties>
-  </rule>
+  <rule ref="category/java/errorprone.xml/NonSerializableClass"/>
 
   <rule ref="category/java/multithreading.xml">
     <exclude name="AvoidUsingVolatile"/>
