--- src/main/java/com/shankhadeepghoshal/springbatch/configs/UtilsConfig.java
@@ -1,22 +1,0 @@
-package com.shankhadeepghoshal.springbatch.configs;
-
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * @author <a href="mailto:shankhadeepghoshal1996@gmail.com">Shankhadeep Ghoshal</a>
- * @since 1.0
- */
-@Configuration
-public class UtilsConfig {
-
-    @Bean
-    public ObjectMapper objectMapper() {
-        final var objectMapper = new ObjectMapper();
-        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-        return objectMapper;
-    }
-}
