+++ src/main/java/com/shankhadeepghoshal/springbatch/personimport/PersonImportJobLauncher.java
@@ -1,6 +1,5 @@
-package com.shankhadeepghoshal.springbatch.launchers;
+package com.shankhadeepghoshal.springbatch.personimport;
 
-import com.shankhadeepghoshal.springbatch.annotations.PurchaseImport;
 import java.util.concurrent.TimeUnit;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
@@ -21,19 +20,18 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
 @Component
-public class PurchaseImportJobLauncher {
-    JobLauncher launcher;
-    @PurchaseImport Job purchaseImportJob;
+public class PersonImportJobLauncher {
 
-    @SneakyThrows
-    @Scheduled(fixedRate = 5L, timeUnit = TimeUnit.SECONDS)
-    public void performJob() {
-        final var jobParameters =
-                new JobParametersBuilder()
-                        .addLong("time", System.currentTimeMillis())
-                        .toJobParameters();
+  JobLauncher launcher;
+  @PersonImport Job personImportJob;
 
-        log.info("Executing Purchase Job....");
-        launcher.run(purchaseImportJob, jobParameters);
-    }
+  @SneakyThrows
+  @Scheduled(fixedRate = 6L, timeUnit = TimeUnit.DAYS)
+  public void performJob() {
+    final var jobParameters =
+        new JobParametersBuilder().addLong("time", System.currentTimeMillis()).toJobParameters();
+
+    log.info("Executing Person Job....");
+    launcher.run(personImportJob, jobParameters);
+  }
 }
