+++ src/main/java/com/shankhadeepghoshal/springbatch/purchaseimport/PurchaseProcessor.java
@@ -1,7 +1,5 @@
-package com.shankhadeepghoshal.springbatch.processors;
+package com.shankhadeepghoshal.springbatch.purchaseimport;
 
-import com.shankhadeepghoshal.springbatch.annotations.PurchaseImport;
-import com.shankhadeepghoshal.springbatch.pojos.Purchase;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.ItemProcessor;
 import org.springframework.stereotype.Component;
@@ -15,9 +13,9 @@ import org.springframework.stereotype.Component;
 @PurchaseImport
 public class PurchaseProcessor implements ItemProcessor<Purchase, Purchase> {
 
-    @Override
-    public Purchase process(Purchase purchase) {
-        log.info("Processing purchase {} ", purchase);
-        return purchase;
-    }
+  @Override
+  public Purchase process(Purchase purchase) {
+    log.info("Processing purchase {} ", purchase);
+    return purchase;
+  }
 }
