+++ src/main/resources/application.properties
@@ -1,10 +1,15 @@
-spring.datasource.url=jdbc:mysql://localhost:3366/people?useSSL=false&allowPublicKeyRetrieval=true
+spring.datasource.url=jdbc:mysql://mysqldb:3306/people?useSSL=false&allowPublicKeyRetrieval=true
 spring.datasource.username=root
 spring.datasource.password=root
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.sql.init.mode=always
 
+server.servlet.context-path=/winner-weekly
+server.port=8080
+
 spring.batch.job.enabled=false
+
+redis.uri=redis://root@redis:6379/
 
 logging.level.sql=trace
 logging.level.web=trace
@@ -17,3 +22,5 @@ job.person=personImportJob
 
 job.purchase.step.first=purchaseStep1
 job.purchase=purchaseImportJob
+
+winner.min-amt=40
