+++ src/main/resources/purchase_record.csv
@@ -19,3 +19,13 @@ custId,amount,buyDate
 8,10.00,2022-12-30
 9,19.00,2022-12-30
 10,10.00,2022-12-30
+1,10.00,2022-12-01
+2,6.00,2022-12-01
+3,8.00,2022-12-01
+4,5.00,2022-12-02
+5,13.00,2022-12-04
+6,5.00,2022-12-03
+7,12.00,2022-12-05
+8,10.00,2022-12-04
+9,21.00,2022-12-02
+10,10.00,2022-11-30
