+++ src/test/java/com/shankhadeepghoshal/springbatch/ApiStepTest.java
@@ -20,19 +20,19 @@ import org.springframework.test.annotation.DirtiesContext;
 @AutoConfigureTestDatabase
 @DirtiesContext
 class ApiStepTest {
-    @Autowired private transient JobLauncherTestUtils jobLauncherTestUtils;
+  @Autowired private transient JobLauncherTestUtils jobLauncherTestUtils;
 
-    @Value("${job.person.step.first}")
-    private transient String stepName;
+  @Value("${job.person.step.first}")
+  private transient String stepName;
 
-    @Test
-    void testStep() {
-        final var jobExecution = jobLauncherTestUtils.launchStep(stepName);
-        final var stepExecution = jobExecution.getStepExecutions().iterator().next();
+  @Test
+  void testStep() {
+    final var jobExecution = jobLauncherTestUtils.launchStep(stepName);
+    final var stepExecution = jobExecution.getStepExecutions().iterator().next();
 
-        Assertions.assertAll(
-                () -> Assertions.assertEquals(BatchStatus.COMPLETED, jobExecution.getStatus()),
-                () -> Assertions.assertEquals(10, stepExecution.getReadCount()),
-                () -> Assertions.assertEquals(10, stepExecution.getWriteCount()));
-    }
+    Assertions.assertAll(
+        () -> Assertions.assertEquals(BatchStatus.COMPLETED, jobExecution.getStatus()),
+        () -> Assertions.assertEquals(10, stepExecution.getReadCount()),
+        () -> Assertions.assertEquals(10, stepExecution.getWriteCount()));
+  }
 }
